frontend: Add indicator if account has rewards to header route
