+++ packages/frontend/src/components/header/HeaderRoutes.tsx
@@ -4,7 +4,9 @@ import Tabs from '@material-ui/core/Tabs'
 import Tab from '@material-ui/core/Tab'
 import { isMainnet } from 'src/config'
 import { useClaim } from 'src/pages/Claim/useClaim'
+import { useHasRewards } from 'src/pages/Rewards/useHasRewards'
 import { makeStyles } from '@material-ui/core/styles'
+import { useTheme } from '@material-ui/core'
 
 const useStyles = makeStyles((theme) => ({
   tabs: {
@@ -19,6 +21,7 @@ const HeaderRoutes: FC = () => {
   const { pathname, search } = useLocation()
   const history = useHistory()
   const styles = useStyles()
+  const theme = useTheme()
 
   const handleChange = (event: ChangeEvent<{}>, newValue: string) => {
     event.preventDefault()
@@ -30,6 +33,7 @@ const HeaderRoutes: FC = () => {
 
   const value = pathname.split('/').slice(0, 2).join('/')
   const { canClaim } = useClaim()
+  const { hasRewards } = useHasRewards()
 
   return (
     <Tabs value={value || '/send'} onChange={handleChange} style={{ width: 'max-content' }} variant="scrollable"
@@ -38,7 +42,7 @@ const HeaderRoutes: FC = () => {
       <Tab label="Pool" value="/pool" />
       <Tab label="Convert" value="/convert" />
       {isMainnet && <Tab label="Stake" value="/stake" />}
-      <Tab label="Airdrop" value="/airdrop" />
+      {/* <Tab label="Airdrop" value="/airdrop" /> */}
       {canClaim && (
         <Tab label="Claim HOP" value="/claim" className="rainbow-animated" style={{
           // background: 'rgba(0, 0, 0, 0) linear-gradient(99.85deg, rgb(179, 46, 255) -18.29%, rgb(242, 164, 152) 109.86%) repeat scroll 0% 0%',
@@ -48,6 +52,12 @@ const HeaderRoutes: FC = () => {
           opacity: 1
         }}/>
       )}
+      <Tab label={<span style={{
+          display: 'inline-block',
+        }}>Rewards {hasRewards && <mark style={{
+        background: 'none',
+        color: theme.palette.primary.main
+      }}>•</mark>}</span>} value="/rewards" />
       {!isMainnet && <Tab label="Faucet" value="/faucet" />}
     </Tabs>
   )
